+++ webapp/__init__.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template
+from flask import Flask
 from flask_login import LoginManager, current_user, login_required
 
 
@@ -6,6 +6,9 @@ from webapp.model import db
 from webapp.user.models import User
 from webapp.user.views import blueprint as user_blueprint
 from webapp.admin.views import blueprint as admin_blueprint
+from webapp.about.views import blueprint as about_blueprint
+from webapp.top.views import blueprint as top_blueprint
+
 
 def create_app():
     app = Flask(__name__)
@@ -17,21 +20,17 @@ def create_app():
     login_manager.login_view = 'user.login'
     app.register_blueprint(user_blueprint)
     app.register_blueprint(admin_blueprint)
+    app.register_blueprint(about_blueprint)
+    app.register_blueprint(top_blueprint)
+
 
     @login_manager.user_loader
     def load_user(user_id):
         return User.query.get(user_id)
 
-    @app.route('/')
-    def index():
-        title = "Главная страница"
-        return render_template('index.html', page_title=title)
+
+
     
-    
-    @app.route('/about')
-    def about():
-        title = "О нас"
-        return render_template('about.html', page_title=title)
     
 
     return app
